pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "tempjacleaner"
version = "0.1.0"
description = "日本語テキスト(特にコード内文字列)の簡易な誤字脱字検出と修正支援ツール"
readme = "README.md"
authors = [
  {name = "Your Name", email = "you@example.com"}
]
license = {file = "LICENSE"}
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Text Processing :: Linguistic",
  "Intended Audience :: Developers"
]
dependencies = []

[project.optional-dependencies]
morph = ["janome>=0.4"]
mecab = [
  "fugashi>=1.3",
  "unidic-lite>=1.0",
]
yaml = ["PyYAML>=6"]
lt = ["language-tool-python>=2.7.1"]
nlp = [
  "spacy>=3.7",
  "ja-ginza>=5.2.0",
]
spell = [
  "rapidfuzz>=3.9",
]

[project.urls]
Homepage = "https://example.com"
Repository = "https://example.com/repo"
Issues = "https://example.com/issues"

[project.scripts]
tempjacleaner = "tempjacleaner.cli:main"

[tool.hatch.build]
include = ["tempjacleaner/**", "README.md", "LICENSE"]
